+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added more decoder integration tests (com -> elpm)
   - added decoder tests (elpm -> lpm)
   - added decoder tests (lsr -> push)
+  - added decoder tests (rcall -> sbic)
 
 - moved sfr_lookup[] table to a separate file. 
 - adjusting readme ;)